1. Quick Start 1. Project skeleton
Made skeleton of project. CMakeLists was set up with gstreamer.
Files and excetuables for it was made.
Project is devided for server and client part.
Server part will be part a device who is sending media (like smartphone).
Client part will be on Raspberry working in console mode. No graphic gui required.